+++ canaille/core/mails.py
@@ -1,11 +1,13 @@
 from flask import current_app
 from flask import url_for
 
-from canaille.app import build_hash
 from canaille.app.i18n import gettext as _
 from canaille.app.mails import logo
 from canaille.app.mails import send_email
 from canaille.app.templating import render_template
+from canaille.backends import Backend
+
+RESET_CODE_LENGTH = 6
 
 
 def send_test_mail(email):
@@ -39,32 +41,45 @@ def send_test_mail(email):
 
 def send_password_reset_mail(user, mail):
     base_url = url_for("core.account.index", _external=True)
-    reset_url = url_for(
-        "core.auth.reset",
-        user=user,
-        hash=build_hash(
-            user.identifier,
-            mail,
-            user.password if user.has_password() else "",
-        ),
-        _external=True,
-    )
+    server_name = current_app.config.get("SERVER_NAME")
     logo_cid, logo_filename, logo_raw = logo()
-
     subject = _("Password reset on {website_name}").format(
         website_name=current_app.config["CANAILLE"]["NAME"]
     )
+
+    reset_token = None
+    reset_url = None
+    reset_code = None
+    if current_app.features.has_trusted_hosts:
+        reset_token = user.generate_url_safe_token()
+        Backend.instance.save(user)
+        reset_url = url_for(
+            "core.auth.reset",
+            user=user,
+            token=reset_token,
+            _external=True,
+        )
+    else:
+        reset_code = user.generate_sms_or_mail_otp(length=RESET_CODE_LENGTH)
+        Backend.instance.save(user)
+
     text_body = render_template(
         "core/mails/reset.txt",
         site_name=current_app.config["CANAILLE"]["NAME"],
         site_url=base_url,
         reset_url=reset_url,
+        server_name=server_name,
+        reset_token=reset_token,
+        reset_code=reset_code,
     )
     html_body = render_template(
         "core/mails/reset.html",
         site_name=current_app.config["CANAILLE"]["NAME"],
         site_url=base_url,
         reset_url=reset_url,
+        server_name=server_name,
+        reset_token=reset_token,
+        reset_code=reset_code,
         logo=f"cid:{logo_cid[1:-1]}" if logo_cid else None,
         title=subject,
     )
@@ -80,14 +95,12 @@ def send_password_reset_mail(user, mail):
 
 def send_password_initialization_mail(user, email):
     base_url = url_for("core.account.index", _external=True)
+    reset_token = user.generate_url_safe_token()
+    Backend.instance.save(user)
     reset_url = url_for(
         "core.auth.reset",
         user=user,
-        hash=build_hash(
-            user.identifier,
-            email,
-            user.password if user.has_password() else "",
-        ),
+        token=reset_token,
         _external=True,
     )
     logo_cid, logo_filename, logo_raw = logo()