+++ tests/core/test_forgotten_password.py
@@ -1,12 +1,15 @@
 import logging
 from unittest import mock
 
+import pytest
+
 
 def test_password_forgotten_disabled(smtpd, testclient, user):
     testclient.app.config["CANAILLE"]["ENABLE_PASSWORD_RECOVERY"] = False
 
     testclient.get("/reset", status=404)
     testclient.get("/reset/user/hash", status=404)
+    testclient.get("/reset-code/user", status=404)
 
     res = testclient.get("/login")
     res.mustcontain(no="Forgotten password")
@@ -156,3 +159,88 @@ def test_password_forgotten_mail_error(SMTP, smtpd, testclient, user):
     res.mustcontain("Send again")
 
     assert len(smtpd.messages) == 0
+
+
+def test_password_forgotten_without_trusted_hosts(
+    smtpd, testclient, user, caplog, backend
+):
+    testclient.app.config["TRUSTED_HOSTS"] = None
+    res = testclient.get("/reset", status=200)
+
+    res.form["login"] = "user"
+    res = res.form.submit(status=302)
+    assert (
+        "success",
+        "A password reset code has been sent at your email address. You should receive "
+        "it within a few minutes.",
+    ) in res.flashes
+    assert (
+        "canaille",
+        logging.SECURITY,
+        "Sending a reset password mail to john@doe.test for user",
+    ) in caplog.record_tuples
+
+    assert len(smtpd.messages) == 1
+
+    res = res.follow(status=200)
+
+    backend.reload(user)
+    code = user.one_time_password
+    email_content = str(smtpd.messages[0].get_payload()[0]).replace("=\n", "")
+    assert code in email_content
+
+    main_form = res.forms[0]
+    main_form["code"] = code
+    res = main_form.submit(status=302)
+    res.follow(status=200)
+
+    assert res.location == f"/reset/{user.user_name}/{code}"
+
+
+def test_password_forgotten_without_trusted_hosts_wrong_code(
+    smtpd, testclient, user, caplog, backend
+):
+    testclient.app.config["TRUSTED_HOSTS"] = None
+    res = testclient.get("/reset", status=200)
+
+    res.form["login"] = "user"
+    res = res.form.submit(status=302)
+    res = res.follow(status=200)
+
+    main_form = res.forms[0]
+    main_form["code"] = "111111"
+    res = main_form.submit(status=200)
+
+    assert ("error", "Invalid code.") in res.flashes
+
+
+def test_password_forgotten_trying_to_access_code_page_with_trusted_hosts_enabled(
+    smtpd, testclient, user
+):
+    testclient.get("/reset-code/user", status=404)
+
+
+def test_password_forgotten_trying_to_access_code_page_when_user_cannot_self_edit(
+    smtpd, testclient, user, backend
+):
+    testclient.app.config["CANAILLE"]["ACL"]["DEFAULT"]["PERMISSIONS"] = []
+    testclient.app.config["TRUSTED_HOSTS"] = None
+    backend.reload(user)
+
+    testclient.app.config["CANAILLE"]["HIDE_INVALID_LOGINS"] = False
+    res = testclient.get("/reset-code/user", status=302)
+    assert (
+        "error",
+        "The user 'John (johnny) Doe' does not have permissions to update their "
+        "password.",
+    ) not in res.flashes
+
+
+def test_password_reset_with_wrong_host(
+    configuration, smtpd, testclient, backend, user
+):
+    res = testclient.get("/reset", status=200)
+    res.form["login"] = "user"
+    # Raises an attribute error but a Security Error is raised before during host validation, causing the attribute error later
+    with pytest.raises(AttributeError):
+        res.form.submit(headers={"Host": "test.test"}, status=400)